Time: 3 ms (35.02%) | Memory: 61.6 MB (48.39%) - LeetSync

+++ 2394-count-subarrays-with-score-less-than-k/count-subarrays-with-score-less-than-k.java
@@ -9,7 +9,7 @@ class Solution {
                 score-=nums[i];
                 i++;
             }
-            res+=(long)j-i+1;
+            res+=j-i+1;
             j++;
         }
         return res;
